[IDE] Move getSelectedOverloadInfo to be a function on Solution
We will need this for solver-based cursor info.
